requirements.txt
# ReqMind Requirements
# RAG System for Requirements Engineering

# LangChain (Modern versions compatible with your installation)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-classic>=1.0.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0

# Vector Store
chromadb>=0.4.22

# Embeddings (Local)
sentence-transformers>=2.3.1

# Document Processing
pypdf>=3.17.4

# LLM API
huggingface-hub>=0.20.3
requests>=2.31.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.12.0
